Move stlite-sharing-communication.ts into StliteSharingIFrame
