+++ packages/sharing-editor/src/StliteSharingIFrame/stlite-sharing-communication.spec.ts
